Added loaders and fixtures
