+++ plugin-server/src/cdp/_tests/fixtures-hogflows.ts
@@ -1,0 +1,75 @@
+import { randomUUID } from 'crypto'
+
+import { HogFlow } from '~/src/schema/hogflow'
+import { insertRow } from '~/tests/helpers/sql'
+
+import { Team } from '../../types'
+import { PostgresRouter } from '../../utils/db/postgres'
+import { UUIDT } from '../../utils/utils'
+import { CyclotronJobInvocationHogFlow, HogFlowInvocationContext } from '../types'
+
+export const createHogFlow = (hogFlow: Partial<HogFlow>) => {
+    const item: HogFlow = {
+        id: randomUUID(),
+        type: 'destination',
+        name: 'Hog Flow',
+        team_id: 1,
+        enabled: true,
+        hog: '',
+        bytecode: [],
+        ...hogFlow,
+    } as HogFlow
+
+    return item
+}
+
+export const insertHogFlow = async (
+    postgres: PostgresRouter,
+    team_id: Team['id'],
+    hogFlow: Partial<HogFlow> = {}
+): Promise<HogFlow> => {
+    // This is only used for testing so we need to override some values
+
+    const res = await insertRow(postgres, 'posthog_hogflow', {
+        ...createHogFlow({
+            ...hogFlow,
+            team_id: team_id,
+        }),
+        description: '',
+        created_at: new Date().toISOString(),
+        updated_at: new Date().toISOString(),
+        created_by_id: 1001,
+        deleted: false,
+    })
+    return res
+}
+
+export const createHogFlowInvocationContext = (
+    data: Partial<HogFlowInvocationContext> = {}
+): HogFlowInvocationContext => {
+    return {
+        ...data,
+    }
+}
+
+export const createExampleHogFlowInvocation = (
+    _hogFlow: Partial<HogFlow> = {},
+    _context: Partial<HogFlowInvocationContext> = {}
+): CyclotronJobInvocationHogFlow => {
+    const hogFlow = createHogFlow(_hogFlow)
+    // Add the source of the trigger to the globals
+
+    const context = createHogFlowInvocationContext(_context)
+
+    return {
+        id: new UUIDT().toString(),
+        state: {
+            ...context,
+        },
+        teamId: hogFlow.team_id,
+        functionId: hogFlow.id,
+        hogFlow,
+        queue: 'hog',
+        queuePriority: 0,
+    }
+}
