+++ plugin-server/src/cdp/services/segment-destination-executor.service.test.ts
@@ -4,13 +4,13 @@ import { defaultConfig } from '~/src/config/config'
 import { fetch, FetchResponse } from '~/src/utils/request'
 import { forSnapshot } from '~/tests/helpers/snapshots'
 
+import { createHogFunction } from '../_tests/fixtures'
 import {
     amplitudeInputs,
     createExampleSegmentInvocation,
-    createHogFunction,
     gameballInputs,
     pipedriveResponse,
-} from '../_tests/fixtures'
+} from '../_tests/fixtures-segment'
 import { SEGMENT_DESTINATIONS_BY_ID } from '../segment/segment-templates'
 import { SegmentDestinationExecutorService } from './segment-destination-executor.service'
